Add pductl tool
Combines the pdu and watchdog packages. It monitors N UDP sockets for
packets and kicks the watchdog if it receives a packet on one.

When a packet is received on an UDP socket, a specific outlet is
turned on (if not already on). After a period of inactivity, the same
outlet is turned off.

I'm using this to switch on a number of audio amplifiers when there is
audio playing and to switch them off if there isn't.